+++ package.json
@@ -16,13 +16,14 @@
     "@mui/icons-material": "^5.14.18",
     "@mui/material": "^5.14.18",
     "@mui/x-date-pickers": "^6.18.1",
+    "@rollup/rollup-darwin-x64": "^4.5.1",
     "@types/react-router-dom": "^5.3.3",
     "dayjs": "^1.11.10",
-    "@types/react-router-dom": "^5.3.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^6.19.0",
     "react-icons": "^4.12.0",
+    "react-router": "^6.20.0",
+    "react-router-dom": "^6.19.0",
     "vite": "^5.0.0"
   },
   "devDependencies": {
